refactor: PackModal -> move photo upload into separate component
